requirements.txt
amqp==1.4.5
anyjson==0.3.3
billiard==3.3.0.17
celery==3.1.11
gnureadline==6.3.3
ipdb==0.8
ipython==2.1.0
kombu==3.0.16
lxml==3.3.5
pymongo==2.7
python-firebase==1.2
pytz==2014.3
redis==2.9.1
requests==1.1.0
wsgiref==0.1.2
